Add JWT Authentication through Spring Security Context to each existing controller class.
Validate the authenticated user in each controller/service method.

Responses are now more consistent throughout the program. Every response now returns a DTO.

Modified REST test files to include the new Authorization Bearer tokens and added instructions.

+++ Backend/src/main/java/backend_api/controller/messaging/ConversationController.java
@@ -5,10 +5,15 @@ import backend_api.DTOs.conversations.ConversationDTO;
 import backend_api.DTOs.conversations.ConversationParticipantResponse;
 import backend_api.DTOs.conversations.ConversationRequest;
 import backend_api.entities.Conversation;
+import backend_api.entities.User;
+import backend_api.enums.ParticipantRole;
 import backend_api.services.ConversationService;
+import backend_api.utils.customexceptions.InvalidConversationRequestException;
+import backend_api.utils.customexceptions.UnauthorizedActionException;
 import backend_api.utils.customexceptions.UserNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,41 +28,50 @@ public class ConversationController {
     }
 
     @PostMapping
-    public ResponseEntity<ConversationDTO> createConversation(@RequestBody ConversationRequest request) {
+    public ResponseEntity<ConversationDTO> createConversation(
+            @RequestBody ConversationRequest request,
+            Authentication authentication) {
+
+        User authUser = (User) authentication.getPrincipal();
+        request.setCreatorId(authUser.getId());
+
         Conversation conversation = conversationService.createAConversation(request);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(ConversationDTO.fromConversationEntity(conversation));
     }
 
     @GetMapping("/{conversationId}")
-    public ResponseEntity<ConversationDTO> getConversationById(@PathVariable Long conversationId) {
+    public ResponseEntity<ConversationDTO> getConversationByIdIfUserIsAParticipant(
+            @PathVariable Long conversationId,
+            Authentication authentication) {
+
+        User authUser = (User) authentication.getPrincipal();
         Conversation conversation = conversationService.getConversationById(conversationId);
-        if (conversation == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
-        }
+        conversationService.validateUserIsParticipant(authUser, conversation);
+
         return ResponseEntity.ok(ConversationDTO.fromConversationEntity(conversation));
     }
 
-    @GetMapping
-    public ResponseEntity<List<ConversationDTO>> getAllConversations() {
-        List<Conversation> conversations = conversationService.getAllConversations();
-        List<ConversationDTO> dtos = conversations.stream()
-                .map(ConversationDTO::fromConversationEntity)
-                .toList();
-        return ResponseEntity.ok(dtos);
-    }
 
-    @GetMapping("/user/{userId}")
-    public ResponseEntity<List<ConversationDTO>> getConversationsByUserId(@PathVariable("userId") Long userId) {
-        List<Conversation> conversations = conversationService.getConversationsByUserId(userId);
+    @GetMapping("/user/me")
+    public ResponseEntity<List<ConversationDTO>> getMyConversations(Authentication authentication) {
+        User authUser = (User) authentication.getPrincipal();
+        List<Conversation> conversations = conversationService.getConversationsByUserId(authUser.getId());
+
         List<ConversationDTO> dtos = conversations.stream()
                 .map(ConversationDTO::fromConversationEntity)
                 .toList();
+
         return ResponseEntity.ok(dtos);
     }
 
     @PutMapping("/{conversationId}/update")
-    public ResponseEntity<?> updateConversation(@PathVariable Long conversationId, @RequestBody ConversationRequest request) {
+    public ResponseEntity<?> updateConversation(@PathVariable Long conversationId,
+                                                @RequestBody ConversationRequest request,
+                                                Authentication authentication) {
+        User authUser = (User) authentication.getPrincipal();
+        conversationService.validateOwnership(authUser, conversationId);
+
         Conversation updatedConversation = conversationService.updateConversation(conversationId, request);
         return ResponseEntity.ok(ConversationDTO.fromConversationEntity(updatedConversation));
     }
@@ -65,7 +79,15 @@ public class ConversationController {
 
     @PutMapping("/{conversationId}/participants/{userId}")
     public ResponseEntity<ConversationParticipantResponse> addUserToConversation(@PathVariable Long conversationId,
-                                                                                 @PathVariable Long userId) {
+                                                                                 @PathVariable Long userId,
+                                                                                 Authentication authentication) {
+
+        User authUser = (User) authentication.getPrincipal();
+        Conversation conversation = conversationService.getConversationById(conversationId);
+        if (!conversation.isParticipant(authUser.getId())) {
+            throw new UnauthorizedActionException("User not a participant of the conversation");
+        }
+
 
         conversationService.addUserToConversation(conversationId, userId);
         ConversationParticipantResponse message = new ConversationParticipantResponse(
@@ -80,7 +102,11 @@ public class ConversationController {
     @DeleteMapping("/{conversationId}/participants/{userId}")
     public ResponseEntity<?> removeUserFromConversation(
             @PathVariable Long conversationId,
-            @PathVariable Long userId) {
+            @PathVariable Long userId,
+            Authentication authentication) {
+
+        User authUser = (User) authentication.getPrincipal();
+        conversationService.validateOwnership(authUser, conversationId);
 
         boolean removed = conversationService.removeUserFromConversation(conversationId, userId);
 
@@ -98,17 +124,21 @@ public class ConversationController {
 
     @PatchMapping("/{conversationId}/leave")
     public ResponseEntity<?> leaveConversation(@PathVariable Long conversationId,
-                                               @RequestParam Long userId) {
-        conversationService.leaveConversation(conversationId, userId);
+                                               Authentication authentication) {
+        User authUser = (User) authentication.getPrincipal();
+        conversationService.leaveConversation(conversationId, authUser.getId());
 
-        return ResponseEntity.ok("User with id " + userId +
-                " has left the conversation " + conversationId);
+        return ResponseEntity.ok("User: " + authUser.getUsername() +
+                " has left the conversation (id): " + conversationId);
     }
 
     @DeleteMapping("/{conversationId}")
     public ResponseEntity<?> deleteConversation(@PathVariable Long conversationId,
-                                                @RequestParam Long requesterId) {
-        conversationService.deleteConversation(conversationId, requesterId);
+                                                Authentication authentication) {
+
+        User authUser = (User) authentication.getPrincipal();
+        conversationService.validateOwnership(authUser, conversationId);
+        conversationService.deleteConversation(conversationId, authUser.getId());
         return ResponseEntity.ok("Conversation deleted successfully");
     }
 }
